refactor(agent-bridge): extract the Codex bridge into skill folders
Summary:
- add root-level `agent-bridge/` and `agent-bridge-codex/` skill
  folders with shared bridge/session/message modules and split Codex
  session, turn, and approval implementations
- keep `SymphonyElixir.Codex.AppServer` and
  `SymphonyElixir.Codex.DynamicTool` as compatibility wrappers while
  switching `AgentRunner` to the generic bridge behaviour
- add generic bridge contract tests plus Codex contract coverage and
  compile the extracted libraries through `mix.exs`

Rationale:
- separate the orchestrator-facing bridge contract from the Codex
  implementation so each folder can evolve toward an independent repo
- move persistent-session idle transport handling behind the bridge
  interface so `AgentRunner` no longer depends on Codex ports
- preserve the repo's existing coverage policy after the extraction by
  ignoring the new bridge implementation modules the same way the old
  monolith was ignored

Tests:
- `mix test test/symphony_elixir/agent_bridge_contract_test.exs test/symphony_elixir/app_server_test.exs test/symphony_elixir/dynamic_tool_test.exs test/symphony_elixir/core_test.exs:2136`
- `make -C elixir all`
