+++ agent-bridge-codex/approval/SKILL.md
@@ -1,0 +1,7 @@
+# Codex Approval Bridge
+
+Owns approval and dynamic-tool handling for Codex turn streams.
+
+- Interpret approval request notifications.
+- Auto-approve when policy allows.
+- Dispatch supported dynamic tools and normalize their responses.
