+++ agent-bridge-codex/approval/lib/symphony_elixir/agent_bridge/codex/approval_policy.ex
@@ -1,0 +1,368 @@
+defmodule SymphonyElixir.AgentBridge.Codex.ApprovalPolicy do
+  @moduledoc false
+
+  alias SymphonyElixir.AgentBridge.Codex.Session
+
+  @non_interactive_tool_input_answer "This is a non-interactive session. Operator input is unavailable."
+
+  @spec handle(port(), String.t(), map(), String.t(), (map() -> term()), map(), function(), boolean()) ::
+          :approved | :approval_required | :input_required | :unhandled
+  def handle(
+        port,
+        "item/commandExecution/requestApproval",
+        %{"id" => id} = payload,
+        payload_string,
+        on_message,
+        metadata,
+        _tool_executor,
+        auto_approve_requests
+      ) do
+    approve_or_require(
+      port,
+      id,
+      "acceptForSession",
+      payload,
+      payload_string,
+      on_message,
+      metadata,
+      auto_approve_requests
+    )
+  end
+
+  def handle(
+        port,
+        "item/tool/call",
+        %{"id" => id, "params" => params} = payload,
+        payload_string,
+        on_message,
+        metadata,
+        tool_executor,
+        _auto_approve_requests
+      ) do
+    tool_name = tool_call_name(params)
+    arguments = tool_call_arguments(params)
+
+    result =
+      tool_name
+      |> tool_executor.(arguments)
+      |> normalize_dynamic_tool_result()
+
+    Session.send_message(port, %{"id" => id, "result" => result})
+
+    event =
+      case result do
+        %{"success" => true} -> :tool_call_completed
+        _ when is_nil(tool_name) -> :unsupported_tool_call
+        _ -> :tool_call_failed
+      end
+
+    Session.emit_message(on_message, event, %{payload: payload, raw: payload_string}, metadata)
+    :approved
+  end
+
+  def handle(
+        port,
+        method,
+        %{"id" => id} = payload,
+        payload_string,
+        on_message,
+        metadata,
+        _tool_executor,
+        auto_approve_requests
+      )
+      when method in ["execCommandApproval", "applyPatchApproval"] do
+    approve_or_require(
+      port,
+      id,
+      "approved_for_session",
+      payload,
+      payload_string,
+      on_message,
+      metadata,
+      auto_approve_requests
+    )
+  end
+
+  def handle(
+        port,
+        "item/fileChange/requestApproval",
+        %{"id" => id} = payload,
+        payload_string,
+        on_message,
+        metadata,
+        _tool_executor,
+        auto_approve_requests
+      ) do
+    approve_or_require(
+      port,
+      id,
+      "acceptForSession",
+      payload,
+      payload_string,
+      on_message,
+      metadata,
+      auto_approve_requests
+    )
+  end
+
+  def handle(
+        port,
+        "item/tool/requestUserInput",
+        %{"id" => id, "params" => params} = payload,
+        payload_string,
+        on_message,
+        metadata,
+        _tool_executor,
+        auto_approve_requests
+      ) do
+    maybe_auto_answer_tool_request_user_input(
+      port,
+      id,
+      params,
+      payload,
+      payload_string,
+      on_message,
+      metadata,
+      auto_approve_requests
+    )
+  end
+
+  def handle(_port, _method, _payload, _payload_string, _on_message, _metadata, _tool_executor, _auto_approve_requests),
+    do: :unhandled
+
+  defp normalize_dynamic_tool_result(%{"success" => success} = result) when is_boolean(success) do
+    output =
+      case Map.get(result, "output") do
+        existing_output when is_binary(existing_output) -> existing_output
+        _ -> dynamic_tool_output(result)
+      end
+
+    content_items =
+      case Map.get(result, "contentItems") do
+        existing_items when is_list(existing_items) -> existing_items
+        _ -> dynamic_tool_content_items(output)
+      end
+
+    result
+    |> Map.put("output", output)
+    |> Map.put("contentItems", content_items)
+  end
+
+  defp normalize_dynamic_tool_result(result) do
+    %{
+      "success" => false,
+      "output" => inspect(result),
+      "contentItems" => dynamic_tool_content_items(inspect(result))
+    }
+  end
+
+  defp dynamic_tool_output(%{"contentItems" => [%{"text" => text} | _]}) when is_binary(text), do: text
+  defp dynamic_tool_output(result), do: Jason.encode!(result, pretty: true)
+
+  defp dynamic_tool_content_items(output) when is_binary(output) do
+    [%{"type" => "inputText", "text" => output}]
+  end
+
+  defp approve_or_require(port, id, decision, payload, payload_string, on_message, metadata, true) do
+    Session.send_message(port, %{"id" => id, "result" => %{"decision" => decision}})
+
+    Session.emit_message(
+      on_message,
+      :approval_auto_approved,
+      %{payload: payload, raw: payload_string, decision: decision},
+      metadata
+    )
+
+    :approved
+  end
+
+  defp approve_or_require(_port, _id, _decision, _payload, _payload_string, _on_message, _metadata, false) do
+    :approval_required
+  end
+
+  defp maybe_auto_answer_tool_request_user_input(
+         port,
+         id,
+         params,
+         payload,
+         payload_string,
+         on_message,
+         metadata,
+         true
+       ) do
+    case tool_request_user_input_approval_answers(params) do
+      {:ok, answers, decision} ->
+        Session.send_message(port, %{"id" => id, "result" => %{"answers" => answers}})
+
+        Session.emit_message(
+          on_message,
+          :approval_auto_approved,
+          %{payload: payload, raw: payload_string, decision: decision},
+          metadata
+        )
+
+        :approved
+
+      :error ->
+        reply_with_non_interactive_tool_input_answer(
+          port,
+          id,
+          params,
+          payload,
+          payload_string,
+          on_message,
+          metadata
+        )
+    end
+  end
+
+  defp maybe_auto_answer_tool_request_user_input(
+         port,
+         id,
+         params,
+         payload,
+         payload_string,
+         on_message,
+         metadata,
+         false
+       ) do
+    reply_with_non_interactive_tool_input_answer(
+      port,
+      id,
+      params,
+      payload,
+      payload_string,
+      on_message,
+      metadata
+    )
+  end
+
+  defp tool_request_user_input_approval_answers(%{"questions" => questions}) when is_list(questions) do
+    answers =
+      Enum.reduce_while(questions, %{}, fn question, acc ->
+        case tool_request_user_input_approval_answer(question) do
+          {:ok, question_id, answer_label} ->
+            {:cont, Map.put(acc, question_id, %{"answers" => [answer_label]})}
+
+          :error ->
+            {:halt, :error}
+        end
+      end)
+
+    case answers do
+      :error -> :error
+      answer_map when map_size(answer_map) > 0 -> {:ok, answer_map, "Approve this Session"}
+      _ -> :error
+    end
+  end
+
+  defp tool_request_user_input_approval_answers(_params), do: :error
+
+  defp reply_with_non_interactive_tool_input_answer(
+         port,
+         id,
+         params,
+         payload,
+         payload_string,
+         on_message,
+         metadata
+       ) do
+    case tool_request_user_input_unavailable_answers(params) do
+      {:ok, answers} ->
+        Session.send_message(port, %{"id" => id, "result" => %{"answers" => answers}})
+
+        Session.emit_message(
+          on_message,
+          :tool_input_auto_answered,
+          %{payload: payload, raw: payload_string, answer: @non_interactive_tool_input_answer},
+          metadata
+        )
+
+        :approved
+
+      :error ->
+        :input_required
+    end
+  end
+
+  defp tool_request_user_input_unavailable_answers(%{"questions" => questions}) when is_list(questions) do
+    answers =
+      Enum.reduce_while(questions, %{}, fn question, acc ->
+        case tool_request_user_input_question_id(question) do
+          {:ok, question_id} ->
+            {:cont, Map.put(acc, question_id, %{"answers" => [@non_interactive_tool_input_answer]})}
+
+          :error ->
+            {:halt, :error}
+        end
+      end)
+
+    case answers do
+      :error -> :error
+      answer_map when map_size(answer_map) > 0 -> {:ok, answer_map}
+      _ -> :error
+    end
+  end
+
+  defp tool_request_user_input_unavailable_answers(_params), do: :error
+
+  defp tool_request_user_input_question_id(%{"id" => question_id}) when is_binary(question_id),
+    do: {:ok, question_id}
+
+  defp tool_request_user_input_question_id(_question), do: :error
+
+  defp tool_request_user_input_approval_answer(%{"id" => question_id, "options" => options})
+       when is_binary(question_id) and is_list(options) do
+    case tool_request_user_input_approval_option_label(options) do
+      nil -> :error
+      answer_label -> {:ok, question_id, answer_label}
+    end
+  end
+
+  defp tool_request_user_input_approval_answer(_question), do: :error
+
+  defp tool_request_user_input_approval_option_label(options) do
+    options
+    |> Enum.map(&tool_request_user_input_option_label/1)
+    |> Enum.reject(&is_nil/1)
+    |> case do
+      labels ->
+        Enum.find(labels, &(&1 == "Approve this Session")) ||
+          Enum.find(labels, &(&1 == "Approve Once")) ||
+          Enum.find(labels, &approval_option_label?/1)
+    end
+  end
+
+  defp tool_request_user_input_option_label(%{"label" => label}) when is_binary(label), do: label
+  defp tool_request_user_input_option_label(_option), do: nil
+
+  defp approval_option_label?(label) when is_binary(label) do
+    normalized_label =
+      label
+      |> String.trim()
+      |> String.downcase()
+
+    String.starts_with?(normalized_label, "approve") or String.starts_with?(normalized_label, "allow")
+  end
+
+  defp tool_call_name(params) when is_map(params) do
+    case Map.get(params, "tool") || Map.get(params, :tool) || Map.get(params, "name") || Map.get(params, :name) do
+      name when is_binary(name) ->
+        case String.trim(name) do
+          "" -> nil
+          trimmed -> trimmed
+        end
+
+      _ ->
+        nil
+    end
+  end
+
+  defp tool_call_name(_params), do: nil
+
+  defp tool_call_arguments(params) when is_map(params) do
+    Map.get(params, "arguments") || Map.get(params, :arguments) || %{}
+  end
+
+  defp tool_call_arguments(_params), do: %{}
+end
