+++ agent-bridge-codex/test/SKILL.md
@@ -1,0 +1,6 @@
+# Codex Bridge Tests
+
+Holds Codex-specific bridge contract helpers and fixtures.
+
+- Reuse the generic `agent-bridge/test` contract macros against the Codex bridge.
+- Keep fake Codex binary setup isolated from orchestrator tests when adding coverage.
