+++ agent-bridge-codex/turn/SKILL.md
@@ -1,0 +1,7 @@
+# Codex Turn Bridge
+
+Owns turn execution for the Codex bridge.
+
+- Start turns on an existing thread.
+- Stream turn messages until completion, timeout, or failure.
+- Emit bridge message events for orchestrator consumers.
